# app/static/swagger.json
{
   "swagger":"2.0",
   "info":{
      "title":"Mi API Flask",
      "version":"1.0"
   },
   "host":"localhost:5000",
   "schemes":[
      "http"
   ],
   "basePath":"/",
   "tags":[
      {
         "name":"Students",
         "description":"Operaciones relacionadas con los estudiantes"
      },
      {
         "name":"Scores",
         "description":"Operaciones relacionadas con los puntajes"
      },
      {
         "name":"Aules",
         "description":"Operaciones relacionadas con las aulas"
      }
   ],
   "paths":{
      "/apps/{mobile_app_id}/students":{
         "get":{
            "tags":[
               "Students"
            ],
            "summary":"Obtener todos los estudiantes de una aplicación móvil específica",
            "parameters":[
               {
                  "name":"mobile_app_id",
                  "in":"path",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{
               "200":{
                  "description":"Lista de estudiantes"
               },
               "404":{
                  "description":"Aplicación móvil no encontrada"
               }
            }
         }
      },
      "/apps/{mobile_app_id}/aules/{aule_id}/students":{
         "get":{
            "tags":[
               "Students"
            ],
            "summary":"Obtener todos los estudiantes de una aula específica",
            "parameters":[
               {
                  "name":"mobile_app_id",
                  "in":"path",
                  "required":true,
                  "type":"string"
               },
               {
                  "name":"aule_id",
                  "in":"path",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{
               "200":{
                  "description":"Lista de estudiantes",
                  "schema":{

                  }
               },
               "404":{
                  "description":"Aplicación móvil o Aula no encontrada"
               }
            }
         },
         "post":{
            "tags":[
               "Students"
            ],
            "summary":"Crear un nuevo estudiante en una aula específica",
            "parameters":[
               {
                  "name":"mobile_app_id",
                  "in":"path",
                  "required":true,
                  "type":"string"
               },
               {
                  "name":"aule_id",
                  "in":"path",
                  "required":true,
                  "type":"string"
               },
               {
                  "name":"student",
                  "in":"body",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Student"
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Estudiante creado"
               },
               "404":{
                  "description":"Aplicación móvil o Aula no encontrada"
               }
            }
         }
      },
      "/apps/{mobile_app_id}/aules/{aule_id}/students/scores":{
         "get":{
            "tags":[
               "Scores"
            ],
            "summary":"Obtener puntajes de estudiantes",
            "description":"Devuelve los puntajes de todos los estudiantes de una aula y aplicación móvil específicas.",
            "parameters":[
               {
                  "name":"mobile_app_id",
                  "in":"query",
                  "required":true,
                  "description":"ID de la aplicación móvil",
                  "type":"string"
               },
               {
                  "name":"aule_id",
                  "in":"query",
                  "required":true,
                  "description":"ID de la aula",
                  "type":"string"
               }
            ],
            "responses":{
               "200":{
                  "description":"Lista de puntajes de los estudiantes",
                  "schema":{
                     "type":"object",
                     "properties":{
                        "results":{
                           "type":"array",
                           "items":{
                              "$ref":"#/definitions/GetScore"
                           }
                        }
                     }
                  }
               },
               "404":{
                  "description":"Aula o aplicación móvil no encontradas"
               },
               "422":{
                  "description":"Parámetros inválidos"
               }
            }
         },
         "post":{
            "tags":[
               "Scores"
            ],
            "summary":"Crear puntaje de estudiante",
            "description":"Crea un nuevo puntaje para un estudiante en una aula y aplicación móvil específicas.",
            "parameters":[
               {
                  "name":"mobile_app_id",
                  "in":"query",
                  "required":true,
                  "description":"ID de la aplicación móvil",
                  "type":"string"
               },
               {
                  "name":"aule_id",
                  "in":"query",
                  "required":true,
                  "description":"ID de la aula",
                  "type":"string"
               },
               {
                  "name":"student_id",
                  "in":"query",
                  "required":true,
                  "description":"ID del estudiante",
                  "type":"integer"
               },
               {
                  "name":"question_id",
                  "in":"query",
                  "required":true,
                  "description":"ID de la pregunta",
                  "type":"string"
               },
               {
                  "name":"miliseconds",
                  "in":"query",
                  "required":true,
                  "description":"Milisegundos que tardó el estudiante en responder la pregunta",
                  "type":"integer"
               },
               {
                  "name":"attempts",
                  "in":"query",
                  "required":true,
                  "description":"Número de intentos que realizó el estudiante para responder la pregunta",
                  "type":"integer"
               }
            ],
            "responses":{
               "201":{
                  "description":"Puntaje creado",
                  "schema":{
                     "$ref":"#/definitions/PostScore"
                  }
               },
               "404":{
                  "description":"Aula o aplicación móvil no encontradas"
               },
               "422":{
                  "description":"Parámetros inválidos"
               }
            }
         }
      },
      "/apps/{mobile_app_id}/aules":{
         "get":{
            "tags":[
               "Aules"
            ],
            "summary":"Obtener todas las aulas de una aplicación móvil específica",
            "parameters":[
               {
                  "name":"mobile_app_id",
                  "in":"path",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{
               "200":{
                  "description":"Lista de aulas",
                  "schema":{
                     "aules":{
                        "type":"array",
                        "items":{
                           "$ref":"#/definitions/Aule"
                        }
                     }
                  }
               },
               "404":{
                  "description":"Aplicación móvil no encontrada"
               }
            }
         },
         "post":{
            "tags":[
               "Aules"
            ],
            "summary":"Crear una nueva aula en una aplicación móvil específica",
            "parameters":[
               {
                  "name":"mobile_app_id",
                  "in":"path",
                  "required":true,
                  "type":"string"
               },
               {
                  "name":"aule",
                  "in":"body",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/Aule"
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Aula creada"
               },
               "404":{
                  "description":"Aplicación móvil no encontrada"
               }
            }
         }
      }
   },
   "definitions":{
      "Student":{
         "type":"object",
         "properties":{
            "first_name":{
               "type":"string"
            },
            "last_name":{
               "type":"string"
            },
            "last_name_2":{
               "type":"string"
            },
            "age":{
               "type":"integer"
            }
         }
      },
      "GetScore":{
         "type":"object",
         "properties":{
            "chapter":{
               "type":"object",
               "properties":{
                  "id":{
                     "type":"string"
                  },
                  "question":{
                     "type":"array",
                     "items":{
                        "type":"object",
                        "properties":{
                           "id":{
                              "type":"string"
                           },
                           "score":{
                              "type":"array",
                              "items":{
                                 "type":"object",
                                 "properties":{
                                    "student_id":{
                                       "type":"integer"
                                    },
                                    "miliseconds":{
                                       "type":"integer"
                                    },
                                    "attempts":{
                                       "type":"integer"
                                    }
                                 }
                              }
                           }
                        }
                     }
                  }
               }
            }
         }
      },
      "PostScore":{
         "type":"object",
         "properties":{
            "student_id":{
               "type":"integer"
            },
            "app_mobile":{
               "type":"object",
               "properties":{
                  "id":{
                     "type":"string"
                  },
                  "chapter":{
                     "type":"object",
                     "properties":{
                        "id":{
                           "type":"string"
                        },
                        "question":{
                           "type":"array",
                           "items":{
                              "type":"object",
                              "properties":{
                                 "id":{
                                    "type":"string"
                                 },
                                 "score":{
                                    "type":"object",
                                    "properties":{
                                       "miliseconds":{
                                          "type":"integer"
                                       },
                                       "attempts":{
                                          "type":"integer"
                                       }
                                    }
                                 }
                              }
                           }
                        }
                     }
                  }
               }
            }
         }
      },
      "Aule":{
         "type":"object",
         "properties":{
            "id":{
               "type":"string"
            },
            "mobile_app_id":{
               "type":"string"
            },
            "name":{
               "type":"string"
            },
            "user_id":{
               "type":"integer"
            },
            "school_id":{
               "type":"integer"
            }
         }
      }
   }
}